+++ src/transpiler/preprocessor/preprocessor.h
@@ -3,6 +3,6 @@
 
 #include <stdio.h>
 
-int prep_readline(char **buf, int *len, FILE *in);
+int prep_readline(char **buf, size_t *len, FILE *in);
 
 #endif
